+++ lib/gen/dartling_library.dart
@@ -49,8 +49,8 @@ String genDartlingLibrary(Model model) {
   sc = '${sc}library ${domain.codeLowerUnderscore}_${model.codeLowerUnderscore}; \n';
   sc = '${sc} \n';
 
-  sc = '${sc}import "dart:convert"; \n';
-  sc = '${sc}import "dart:math"; \n';
+  sc = '${sc}//import "dart:convert"; \n';
+  sc = '${sc}//import "dart:math"; \n';
   sc = '${sc} \n';
 
   sc = '${sc}import "package:dartling/dartling.dart"; \n';